[ncicore] Fixed credit management. JB#48413
NFCForum-TS-NCI-1.0

4.4.4 Connection Credit Management

...
If there are not enough credits to send the whole Data Message,
the DH is allowed to send as many Data Packets as number of credits
available.

When the DH wants to send a Data Packet on a Logical Connection and
flow control is enabled, the DH SHALL check that the NFCC_Credits_Avail
variable for that connection is greater than 0. If so, the DH SHALL
reduce the NFCC_Credits_Avail by 1 and transfer the Data Packet over
the Logical Connection. The DH SHALL NOT send any Data Packet on a
connection when the corresponding NFCC_Credits_Avail is 0.